requirements.txt
alembic==1.8.1
autopep8==1.6.0        
black==22.6.0
Bootstrap-Flask==2.1.0 
click==8.1.3
colorama==0.4.5        
dominate==2.7.0        
Flask==2.2.2
Flask-Bootstrap4==4.0.2
Flask-Migrate==3.1.0   
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0  
itsdangerous==2.1.2    
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1      
mypy-extensions==0.4.3 
pathspec==0.9.0        
platformdirs==2.5.2    
pycodestyle==2.9.0     
SQLAlchemy==1.4.42     
toml==0.10.2
tomli==2.0.1
visitor==0.1.3
Werkzeug==2.2.2        
WTForms==3.0.1